# Add AES CBC psk encryption