--- a/n2n.h
+++ b/n2n.h
@@ -186,6 +186,7 @@ typedef char n2n_sn_name_t[N2N_EDGE_SN_HOST_SIZE];
 
 struct n2n_edge {
   int                 daemon;                 /**< Non-zero if edge should detach and run in the background. */
+  int                 preferred_aes;          /**< Non-zero if AES is the preferred encryption meothd. */
   uint8_t             re_resolve_supernode_ip;
 
   n2n_sock_t          supernode;
@@ -337,7 +338,9 @@ void set_peer_operational(n2n_edge_t * eee,
 			  const n2n_mac_t mac,
 			  const n2n_sock_t * peer);
 const char * supernode_ip(const n2n_edge_t * eee);
-int edge_init_twofish(n2n_edge_t * eee, uint8_t *encrypt_pwd,
+int edge_init_twofish_psk(n2n_edge_t * eee, uint8_t *encrypt_pwd,
+		      uint32_t encrypt_pwd_len);
+int edge_init_aes_psk(n2n_edge_t * eee, uint8_t *encrypt_pwd,
 		      uint32_t encrypt_pwd_len);
 int run_edge_loop(n2n_edge_t * eee, int *keep_running);
 void edge_term(n2n_edge_t * eee);