Adds AES encryption support for pre shared key mode. Previously AES was only enabled when a keyfile was passed with the `-K` (uppercase) option. With this pull request it's possible to use AES with `-k` (lowercase) too, by adding the `-A` option. Encription in use can be verified with using the `-v` option.

Without `-A` (default):
`28/Jan/2019 01:08:42 [transform_tf.c:132] encode_twofish 98 with SA 0.`

With `-A`:
`28/Jan/2019 01:13:09 [transform_aes.c:162] encode_aes 98 with SA 0.`

AES should become the preferred encryption method in the future as, contrary to twofish, it uses a standard system library. Here are some benchmarks on my system (`./benchmark`):

```
[transop_null] 10000 times: (6356 -> 635 nsec each) 1548636359.095106 -> 1548636359.101462.
[transop_twofish] 10000 times: (494206 -> 49420 nsec each) 1548636359.101502 -> 1548636359.595708.
[transop_aes] 10000 times: (63927 -> 6392 nsec each) 1548636359.595725 -> 1548636359.659652.
````

Note: system with AES hardware accelleration