@lucaderi 2019-01-28
The main issue I see for making it mainstream is to enable interoperation with edge instances that have not been started with -A