@Logan008 2019-01-30
Wow, this is awesome!  Thank you!!

I see no problem for the mainstream, as it probably only is useful in conjunction with -k; so it could be considered as some kind of key related parameter which needs to be given manually anyway when invoking edge using -k.  It just needs to be communicated together with the psk which needs to be shared somehow anyway on some occasion.

Does it also work using N2N_KEY environment?  I need to check it out immediately.