@emanuele-f 2019-01-31
Hi, it works for me with the N2N_KEY env. N2N_KEY is equivalent to `-k`