@Logan008 2019-01-31
Oh, my bad!  You are so right, of course.  I should have watched better the path to the executable file...   Maybe I am getting too old for this.  I am sorry and embarrassed.

Anyway, "-A" is something I dreamt of!  Wonderful work and thank you again!

A real quick benchmark here (some older i7-2860QM including cpu hw aes flag) delivers the following output:
```
[transop_null] 10000 times: (2822 -> 282 nsec each) 1548964976.065853 -> 1548964976.068675.
[transop_twofish] 10000 times: (376056 -> 37605 nsec each) 1548964976.068716 -> 1548964976.444772.
[transop_aes] 10000 times: (52552 -> 5255 nsec each) 1548964976.444788 -> 1548964976.497340.
```

I guess, real life throughput is much more limited by bandwitdth.