@emanuele-f 2019-02-01
7x speed increase is not bad :) Next step to improve n2n is to fix the connection issues like #28 . Regarding encryption, we also have to evaluate if adding some salt or hash on the raw passphrase before feeding it into aes/twofish is better from a security point.