@Logan008 2019-02-05
A hash on the key before feeding it into the AES or TF would be a good thing to do, as hashed values usually spread the bits broader and look more pseudo-random as an ASCII coded password ever could. It makes more unpredictable use of the available key space.

A salt would provide additional security in situations where the password hash (or, in the case mentioned above: a hash of the hash) should be stored (together with the salt) for verification purposes, i.e. authentication. So you probably do not need salt for now.

SHA256 or similar should do for AES with 256 bit key size.

By the way, wouldn't a 512 bit key (and thus SHA512) be an even more secure choice with a view to a block size of 256 bit? There are 256! permutations for blocks of 256 bits, by far not each of them could be named by a 256 bit key (2 ^ 256 << 256!) ; 2 ^ 512 gets at least a bit closer...