@emanuele-f 2019-02-19
@Logan008 if possible please open a new feature request for the use of an hash function and explain with some examples why using SHA512 would be a better choice than SHA256, thank you